pco_api v0.0.1 PcoApi.Query

A set of functions to assist in composing queries to the API.

Functions

where(param)

Creates a list of where params to pass onto the api call. Returns a List of Tuples.

Examples

iex> PcoApi.Query.where(first_name: "Geoffrey")
[{"where[first_name]", "Geoffrey"}]

iex> PcoApi.Query.where(first_name: "Geoffrey") |> PcoApi.Query.where(last_name: "Lessel")
[{"where[last_name]", "Lessel"}, {"where[first_name]", "Geoffrey"}]

where(params, param)